refactor: keep project helpers in handlers
