+++ packages/cli/src/handlers/project-remove.ts
@@ -1,8 +1,14 @@
+import { realpath } from "node:fs/promises";
+import { isAbsolute, resolve } from "node:path";
 import { parseArgs } from "node:util";
-import { ServeStateStore, type ProjectRecord } from "@phantompane/state";
+import { getGitRoot } from "@phantompane/git";
+import {
+  ServeStateStore,
+  type ProjectRecord,
+  type ServeState,
+} from "@phantompane/state";
 import { exitCodes, exitWithError, exitWithSuccess } from "../errors.ts";
 import { output } from "../output.ts";
-import { findProject, hasBlockingProjectChat } from "./project-utils.ts";
 
 export async function projectRemoveHandler(args: string[] = []): Promise<void> {
   const { positionals } = parseArgs({
@@ -96,4 +102,53 @@ export async function projectRemoveHandler(args: string[] = []): Promise<void> {
   );
 
   exitWithSuccess();
+}
+
+async function findProject(
+  state: ServeState,
+  identifier: string,
+): Promise<ProjectRecord | null> {
+  const directMatches = state.projects.filter(
+    (project) =>
+      project.id === identifier ||
+      project.name === identifier ||
+      project.rootPath === identifier,
+  );
+
+  if (directMatches.length === 1) {
+    return directMatches[0]!;
+  }
+
+  if (directMatches.length > 1) {
+    throw new Error(`Project '${identifier}' is ambiguous`);
+  }
+
+  try {
+    const rootPath = await resolveProjectRootPath(identifier);
+    return (
+      state.projects.find((project) => project.rootPath === rootPath) ?? null
+    );
+  } catch {
+    return null;
+  }
+}
+
+async function resolveProjectRootPath(path: string): Promise<string> {
+  const absolutePath = isAbsolute(path) ? path : resolve(path);
+  const resolvedPath = await realpath(absolutePath);
+  return await getGitRoot({ cwd: resolvedPath });
+}
+
+function hasBlockingProjectChat(state: ServeState, projectId: string): boolean {
+  const queuedChatIds = new Set(
+    state.queuedMessages.map((message) => message.chatId),
+  );
+  return state.chats.some(
+    (chat) =>
+      chat.projectId === projectId &&
+      (Boolean(chat.activeTurnId) ||
+        chat.status === "running" ||
+        chat.status === "waitingForApproval" ||
+        queuedChatIds.has(chat.id)),
+  );
 }
